+++ build_runner_core/lib/src/asset_graph/optional_output_tracker.dart
@@ -73,7 +73,9 @@ class OptionalOutputTracker {
     return _checkedOutputs.putIfAbsent(
       output,
       () =>
-          node.outputs.any((o) => isRequired(o, currentlyChecking)) ||
+          (_assetGraph.computeOutputs()[node.id] ?? <AssetId>{}).any(
+            (o) => isRequired(o, currentlyChecking),
+          ) ||
           _assetGraph
               .outputsForPhase(output.package, nodeConfiguration.phaseNumber)
               .where(